Perform linewise paste, if regtype 'V' is detected
Previously, to perform a linewise paste at cursors it was necessary
to run the normal command 'p' or 'P'. Now, if the register from which
you're pasting is of type V and there is no VM register (so that vim
register is going to be used instead), a linewise paste is performed
with the mappings p/P (no need for normal command).

Additionally: default VM_leader is always '\\'.