Smart case change
When change is started with 'gc' rather than 'c', smart case change will
be used. In this mode, at each cursor, the inserted text will be
capitalized or uppercased, if the replaced text was.

Moreover, for more complex transformations of the inserted text, one may
run the command VMSmartChange that accepts as argument the name of
a function, eg. MyTextTransform(cursor, text), where
  - a:cursor is the handled cursor
  - a:text is the text inserted by the main cursor

:VMSmartChange MyTextTransform

fun! MyTextTransform(cursor, txt) abort
    " the active cursor isn't affected, text is entered as typed
    try
        let original_text = b:VM_Selection.Vars.changed_text[a:cursor.index]
        if match(original_text, '\u') >= 0 && match(original_text, '\U') < 0
            return toupper(a:txt)
        elseif match(original_text, '\u') == 0
            return toupper(a:txt[:0]) . a:txt[1:]
        else
            return a:txt
        endif
    catch
        return a:txt
    endtry
endfun