Check column mismatch at InsertLeave
At every text change, the last known edited column is stored, then at
InsertLeave it is confronted with the column at which insert mode was
exited. They can be different if there was abbreviation expansion or
popup completion, in this case update the text one more time.

Fixes unexpanded abbreviations on InsertLeave with vim (in nvim they
were already working). CompleteDone was already implemented so no change
in that regard.