fix: surround mapping (#272)
Alternative to #271, storing the mapping in a different variable for consistency.

The <plug>(VM-Surround) did nothing good, since the key was hard-coded
for use with vim-surround. Now it will use the keys set ing:VM_maps['Surround'].
Using something else than vim-surround breaks surrounding with tags.
From cursor mode `ys` is still hard-coded. Not much I can do about it
for now (not worth to rethink all the mappings system just for this).